requirements.txt
Flask==3.0.3

# Database - SQLAlchemy + PostgreSQL
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Databricks SDK (for OAuth token generation and agent queries)
# IMPORTANT: Version 0.56.0+ required for database.generate_database_credential()
databricks-sdk>=0.56.0

python-dotenv>=1.0.0